+++ apps/func-migration-v3-import/src/index.ts
@@ -20,9 +20,11 @@ const blobTrigger: StorageBlobHandler = async function (
 
     const content = data.toString()
 
+    context.log(context.triggerMetadata?.blobTrigger)
+
     const newUserId = context.triggerMetadata?.blobTrigger
-      ?.split('/')[2]
-      .split('_')[0]
+      ?.split('/')
+      [process.env.NODE_ENV === 'development' ? 1 : 2].split('_')[0]
 
     const parsedContent = JSON.parse(content)
 
@@ -44,7 +46,8 @@ const blobTrigger: StorageBlobHandler = async function (
 
     sendTeamsNotifications(
       'func/migration-v3-import',
-      `Started import of KlickerV2 data for user '${user.email}'`
+      `Started import of KlickerV2 data for user '${user.email}'`,
+      context
     )
 
     // keep information in memory for more efficient lookup
@@ -54,10 +57,16 @@ const blobTrigger: StorageBlobHandler = async function (
     let mappedQuestionInstancesIds: Record<string, number> = {}
     let mappedSessionIds: Record<string, string> = {}
 
-    mappedFileURLs = await migrateFiles(parsedContent.files, context, user)
+    mappedFileURLs = await migrateFiles(
+      prisma,
+      parsedContent.files,
+      context,
+      user
+    )
+
     // context.log('mappedFileURLs: ', mappedFileURLs)
 
-    const batchSize = 50
+    const batchSize = 10
 
     const tags = parsedContent.tags
     mappedTags = await importTags(prisma, tags, user, batchSize, context)
@@ -97,18 +106,20 @@ const blobTrigger: StorageBlobHandler = async function (
     )
     // context.log('mappedSessionIds: ', mappedSessionIds)
 
-    sendTeamsNotifications(
+    await sendTeamsNotifications(
       'func/migration-v3-import',
-      `Successful migration for user '${user.email}'`
+      `Successful migration for user '${user.email}'`,
+      context
     )
-    sendEmailMigrationNotification(user.email, true)
+    await sendEmailMigrationNotification(user.email, true, context)
   } catch (e) {
     context.error('Something went wrong while importing data: ', e)
-    sendTeamsNotifications(
+    await sendTeamsNotifications(
       'func/migration-v3-import',
-      `Migration of KlickerV2 data failed. Error: ${e.message}`
+      `Migration of KlickerV2 data failed. Error: ${e.message}`,
+      context
     )
-    sendEmailMigrationNotification(email, false)
+    await sendEmailMigrationNotification(email, false, context)
     throw new Error('Something went wrong while importing data')
   } finally {
     await closeLegacyConnection()
@@ -120,6 +131,6 @@ export default blobTrigger
 
 app.storageBlob('MigrationV3Import', {
   connection: 'MIGRATION_BLOB_IMPORT_CONNECTION_STRING',
-  path: process.env.MIGRATION_BLOB_STORAGE_PATH as string,
+  path: 'exports',
   handler: blobTrigger,
 })