+++ packages/i18n/messages/en.ts
@@ -7,26 +7,31 @@ export default {
       ftPlaceholder: 'Your answer...',
     },
     SC: {
+      short: 'SC',
       typeLabel: 'Single Choice (SC)',
       text: 'Please select a single option.',
       richtext: 'Please select a <b>single</b> option.',
     },
     MC: {
+      short: 'MC',
       typeLabel: 'Multiple Choice (MC)',
       text: 'Please select one or more options.',
       richtext: 'Please select <b>one or more</b> options.',
     },
     KPRIM: {
+      short: 'KP',
       typeLabel: 'Kprim (KP)',
       text: 'Evaluate the statements for correctness.',
       richtext: 'Evaluate the statements for <b>correctness</b>.',
     },
     FREE_TEXT: {
+      short: 'FT',
       typeLabel: 'Free Text (FT)',
       text: 'Please enter your answer.',
       richtext: 'Please enter your <b>answer</b>.',
     },
     NUMERICAL: {
+      short: 'NR',
       typeLabel: 'Numerical (NR)',
       text: 'Please enter a number.',
       richtext: 'Please enter a <b>number</b>.',
@@ -596,7 +601,7 @@ If multipliers are activated on the activities, these multipliers are applied to
       pageName: 'Migrate Data from KlickerV2',
       step1Title: 'Step 1: Request Migration Token',
       step1Description:
-        'To migrate your old Klicker account to KlickerV3, provide the email linked to your old account. After submitting the email, you will receive a migration token to proceed. Please ensure you have access to this email. If not, consider resetting your password.',
+        'To migrate your old account to KlickerUZH v3.0, provide the e-mail linked to your old account. After submitting the e-mail, you will receive a link to proceed. If you do not receive a message on the provided e-mail within the next 5 minutes, please check your spam folder and try starting the migration process again. Please ensure you have access to your old e-mail inbox.',
       requestMigrationToken: 'Request Migration Token',
       step2Title: 'Step 2: Insert Migration Link',
       step2Description:
@@ -661,7 +666,6 @@ If multipliers are activated on the activities, these multipliers are applied to
       explanationTooltip:
         'Enter a generic explanation of your question here, which will be displayed to students in practice quizzes and microlearning regardless of their answer as an explanation of the correct solution.',
       explanationPlaceholder: 'Enter your explanation here...',
-      attachments: 'Attachments',
       answerOptions: 'Answer options',
       answerOption: 'Answer option',
       answerOptionsTooltip:
@@ -687,6 +691,10 @@ If multipliers are activated on the activities, these multipliers are applied to
       editElement: 'Edit {element}',
       cancelCreation: 'Cancel creation',
       cancelEditing: 'Cancel editing',
+      mediaLibrary: 'Media Library',
+      uploadImageHeader: 'Upload Media',
+      uploadImageDescription:
+        'Drag an image here to upload or click to open file explorer.',
     },
     sessionForms: {
       sessionName: 'Please enter a name for your quiz.',
@@ -815,8 +823,6 @@ If multipliers are activated on the activities, these multipliers are applied to
         'Please ensure that the following errors in the form are resolved before saving the question:',
       questionName: 'Please enter a name for the question.',
       questionContent: 'Please add some content to your question.',
-      attachmentURL: 'Please enter a valid URL for the attachment.',
-      attachmentName: 'Please enter a name for the attachment.',
       answerContent: 'Please add some content to your answer option.',
       feedbackContent: 'Please add some content to your answer feedback.',
       SCAnswersCorrect: