+++ packages/next-config/index.js
@@ -26,7 +26,13 @@ function getNextBaseConfig({ S3_HOSTNAME, S3_PATHNAME, NODE_ENV }) {
       },
     },
     images: {
-      domains: ['127.0.0.1', 'tc-klicker-prod.s3.amazonaws.com', S3_HOSTNAME, "https://klickermigration.blob.core.windows.net"],
+      domains: [
+        '127.0.0.1',
+        'tc-klicker-prod.s3.amazonaws.com',
+        'klickeruzhdevimages.blob.core.windows.net',
+        'klickeruzhprodimages.blob.core.windows.net',
+        S3_HOSTNAME,
+      ],
       remotePatterns: [
         {
           protocol: 'https',
@@ -36,15 +42,21 @@ function getNextBaseConfig({ S3_HOSTNAME, S3_PATHNAME, NODE_ENV }) {
         },
         {
           protocol: 'https',
-          hostname: S3_HOSTNAME,
+          hostname: 'klickeruzhdevimages.blob.core.windows.net',
           port: '443',
-          pathname: S3_PATHNAME,
+          pathname: '/**',
         },
         {
           protocol: 'https',
-          hostname: 'https://klickermigration.blob.core.windows.net',
+          hostname: 'klickeruzhprodimages.blob.core.windows.net',
           port: '443',
-          pathname: '/images/**',
+          pathname: '/**',
+        },
+        {
+          protocol: 'https',
+          hostname: S3_HOSTNAME,
+          port: '443',
+          pathname: S3_PATHNAME,
         },
       ],
     },